(Gapi-3) Refactor API.
