+++ exploiter.js
@@ -1,4 +1,3 @@
-const EventEmitter = require('events')
 const fetch = require('node-fetch');
 const credentials = require('./credentials.json')
 const baseURL =  'https://api.github.com/';
